requirements.txt
# Sistema experto + red bayesiana
experta==1.9.4
pgmpy==0.1.18

# Requisitos binarios compatibles
numpy==1.23.5
pandas==1.5.3

# Framework web y pruebas
flask==2.2.5
pytest==7.4.0

# Telegram bot
python-telegram-bot==20.7

# Servidor de producción
gunicorn

python-dotenv>=1.0.0
matplotlib
psutil
kagglehub
